Refactored capture and save
